+++ src/lib/kotlin/slatekit-integration/src/main/kotlin/slatekit/integration/jobs/APIWorker.kt
@@ -6,7 +6,7 @@ import slatekit.apis.core.Requests
 import slatekit.common.Identity
 import slatekit.common.Sources
 import slatekit.jobs.*
-import slatekit.jobs.slatekit.jobs.Worker
+import slatekit.jobs.Worker
 import slatekit.results.Failure
 import slatekit.results.Success
 import slatekit.results.Try
